Let `features_id` be str or a subclass of `msaf.features.Features`
For collection mode, it's not possible to add a new feature to the
feature_registry dict from the outside of MSAF. This is because the
feature_registry is cleared out per joblib worker.

By allowing
```
class MyFeature(msaf.features.Features):
    ...

msaf.run.process(...feature=MyFeature)
```
it becomes possible to add new features and run with collection mode.